Changes as of comments from 6 Mar 2024, from PR 14
