+++ ndsl/checkpointer/base.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class Checkpointer(abc.ABC):
+    @abc.abstractmethod
+    def __call__(self, savepoint_name, **kwargs):
+        ...
