+++ ndsl/checkpointer/null.py
@@ -1,4 +1,4 @@
-from ndsl.typing import Checkpointer
+from ndsl.checkpointer.base import Checkpointer
 
 
 class NullCheckpointer(Checkpointer):
