+++ ndsl/dsl/caches/cache_location.py
@@ -1,5 +1,5 @@
+from ndsl.comm.partitioner import Partitioner
 from ndsl.dsl.caches.codepath import FV3CodePath
-from ndsl.typing import Partitioner
 
 
 def identify_code_path(
