+++ ndsl/dsl/stencil.py
@@ -23,6 +23,7 @@ from gt4py.cartesian import gtscript
 from gt4py.cartesian.gtc.passes.oir_pipeline import DefaultPipeline, OirPipeline
 
 from ndsl.comm.comm_abc import Comm
+from ndsl.comm.communicator import Communicator
 from ndsl.comm.decomposition import block_waiting_for_compilation, unblock_waiting_tiles
 from ndsl.comm.mpi import MPI
 from ndsl.constants import X_DIM, X_DIMS, Y_DIM, Y_DIMS, Z_DIM, Z_DIMS
@@ -34,7 +35,6 @@ from ndsl.quantity import Quantity
 
 # from ndsl import testing
 from ndsl.testing import comparison
-from ndsl.typing import Communicator
 
 
 try:
