+++ ndsl/dsl/stencil_config.py
@@ -5,10 +5,11 @@ from typing import Any, Callable, Dict, Hashable, Iterable, Optional, Sequence, 
 
 from gt4py.cartesian.gtc.passes.oir_pipeline import DefaultPipeline, OirPipeline
 
+from ndsl.comm.communicator import Communicator
 from ndsl.comm.decomposition import determine_rank_is_compiling, set_distributed_caches
+from ndsl.comm.partitioner import Partitioner
 from ndsl.dsl.dace.dace_config import DaceConfig, DaCeOrchestration
 from ndsl.dsl.gt4py_utils import is_gpu_backend
-from ndsl.typing import Communicator, Partitioner
 
 
 class RunMode(enum.Enum):
