+++ ndsl/grid/generation.py
@@ -5,6 +5,7 @@ from typing import Tuple
 
 import numpy as np
 
+from ndsl.comm.communicator import Communicator
 from ndsl.constants import (
     N_HALO_DEFAULT,
     PI,
@@ -58,7 +59,6 @@ from ndsl.stencils.corners import (
     fill_corners_cgrid,
     fill_corners_dgrid,
 )
-from ndsl.typing import Communicator
 
 
 # TODO: when every environment in python3.8, remove
