+++ ndsl/monitor/zarr_monitor.py
@@ -4,13 +4,12 @@ from typing import List, Tuple, Union
 import cftime
 
 import ndsl.constants as constants
-from ndsl.comm.partitioner import subtile_slice
+from ndsl.comm.partitioner import Partitioner, subtile_slice
 from ndsl.logging import ndsl_log
 from ndsl.monitor.convert import to_numpy
 from ndsl.optional_imports import cupy
 from ndsl.optional_imports import xarray as xr
 from ndsl.optional_imports import zarr
-from ndsl.typing import Partitioner
 from ndsl.utils import list_by_dims
