+++ ndsl/stencils/c2l_ord.py
@@ -8,13 +8,13 @@ from gt4py.cartesian.gtscript import (
 )
 
 import ndsl.dsl.gt4py_utils as utils
+from ndsl.comm.communicator import Communicator
 from ndsl.constants import X_DIM, X_INTERFACE_DIM, Y_DIM, Y_INTERFACE_DIM, Z_DIM
 from ndsl.dsl.dace.wrapped_halo_exchange import WrappedHaloUpdater
 from ndsl.dsl.stencil import StencilFactory
 from ndsl.dsl.typing import Float, FloatField, FloatFieldIJ
 from ndsl.grid.helper import GridData
 from ndsl.initialization.allocator import QuantityFactory
-from ndsl.typing import Communicator
 
 
 A1 = 0.5625
