+++ ndsl/stencils/testing/conftest.py
@@ -8,7 +8,11 @@ import xarray as xr
 import yaml
 
 import ndsl.dsl
-from ndsl.comm.communicator import CubedSphereCommunicator, TileCommunicator
+from ndsl.comm.communicator import (
+    Communicator,
+    CubedSphereCommunicator,
+    TileCommunicator,
+)
 from ndsl.comm.mpi import MPI
 from ndsl.comm.partitioner import CubedSpherePartitioner, TilePartitioner
 from ndsl.dsl.dace.dace_config import DaceConfig
@@ -16,7 +20,6 @@ from ndsl.namelist import Namelist
 from ndsl.stencils.testing.parallel_translate import ParallelTranslate
 from ndsl.stencils.testing.savepoint import SavepointCase, dataset_to_dict
 from ndsl.stencils.testing.translate import TranslateGrid
-from ndsl.typing import Communicator
 
 
 @pytest.fixture()
