+++ tests/dsl/test_stencil_factory.py
@@ -3,7 +3,6 @@ import pytest
 from gt4py.cartesian.gtscript import PARALLEL, computation, horizontal, interval, region
 
 from ndsl import (
-    CompareToNumpyStencil,
     CompilationConfig,
     DaceConfig,
     FrozenStencil,
@@ -13,7 +12,7 @@ from ndsl import (
 )
 from ndsl.constants import X_DIM, Y_DIM, Z_DIM
 from ndsl.dsl.gt4py_utils import make_storage_from_shape
-from ndsl.dsl.stencil import get_stencils_with_varied_bounds
+from ndsl.dsl.stencil import CompareToNumpyStencil, get_stencils_with_varied_bounds
 from ndsl.dsl.typing import FloatField
