+++ tests/test_halo_data_transformer.py
@@ -4,13 +4,8 @@ from typing import Tuple
 import numpy as np
 import pytest
 
-from ndsl import (
-    Buffer,
-    HaloDataTransformer,
-    HaloExchangeSpec,
-    Quantity,
-    QuantityHaloSpec,
-)
+from ndsl import HaloExchangeSpec, Quantity
+from ndsl.buffer import Buffer
 from ndsl.comm import _boundary_utils
 from ndsl.constants import (
     EAST,
@@ -28,7 +23,9 @@ from ndsl.constants import (
     Z_DIM,
     Z_INTERFACE_DIM,
 )
+from ndsl.halo import HaloDataTransformer
 from ndsl.halo.rotate import rotate_scalar_data, rotate_vector_data
+from ndsl.quantity import QuantityHaloSpec
 
 
 @pytest.fixture
